pipeline/requirements.txt
osmium
psycopg2-binary
shapely
numpy
rasterio
requests
python-dotenv
